chore: move price info, trading view widget components into src/components/common
